Restructured the entire codebase
