--- Java/NeuralNetwork.Java
@@ -1,14 +1,0 @@
-Public Class NeuralNetwork {
-    static float[] input;
-    static float[] input1;
-    static float[] input2;
-    
-    ...
-
-    public static void main(String[] args)
-
-    
-    ....
-
-
-}
