+++ tests/__init__.py
@@ -1,0 +1,2 @@
+from .activations import TestActivationFunctions
+from .dense import DenseTest
